test(api): bdd migrate zero-chat-threads-artifacts
